장부 레포트 화면 데이터 연동 및 월별 조회 기능 구현
퍼센트 값 소수점 제거(내림 처리)로 인한 Int로 설정

+++ feature/report/src/main/java/com/moneymong/moneymong/report/model/MemberReport.kt
@@ -4,8 +4,8 @@ data class MemberReport(
     val name: String,
     val income: Long,
     val expense: Long,
-    val incomePercent: Double,
-    val expensePercent: Double
+    val incomePercent: Int,
+    val expensePercent: Int
 )
 
 internal val mockMemberReports: List<MemberReport> = listOf(
@@ -13,14 +13,14 @@ internal val mockMemberReports: List<MemberReport> = listOf(
         name = "장희직",
         income = 5000L,
         expense = 5000L,
-        incomePercent = 100.0,
-        expensePercent = 50.0
+        incomePercent = 100,
+        expensePercent = 50
     ),
     MemberReport(
         name = "김희직",
         income = 45000L,
         expense = 5000L,
-        incomePercent = 90.0,
-        expensePercent = 50.0
+        incomePercent = 90,
+        expensePercent = 50
     ),
 )
