+++ feature/report/src/main/java/com/moneymong/moneymong/report/model/MonthlyReport.kt
@@ -3,6 +3,6 @@ package com.moneymong.moneymong.report.model
 data class MonthlyReport(
     val income: Long,
     val expense: Long,
-    val incomePercent: Double,
-    val expensePercent: Double
+    val incomePercent: Int,
+    val expensePercent: Int
 )
